requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==5.0.4
django-ratelimit==4.1.0
executing==2.0.1
fastapi==0.110.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.7
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
modal==0.62.78
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
safetensors==0.4.2
sigtools==4.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
synchronicity==0.6.7
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
typer==0.9.4
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
yarl==1.9.4